perf: Standardize document naming
