+++ src-tauri/src/commands/asset_unfavorite.rs
@@ -1,10 +1,11 @@
-use crate::commands::auth_login::ensure_fresh_token;
-use crate::service::favorite::FavoriteService;
+use crate::commands::auth::ensure_fresh_token;
+use crate::service::asset_favorite::FavoriteService;
+use log::info;
 use serde_json::json;
 use tauri::{AppHandle, Emitter};
 
 #[tauri::command]
-pub async fn set_favorite(
+pub async fn unfavorite(
     app: AppHandle,
     site: String,
     bearer_token: String,
@@ -19,18 +20,20 @@ pub async fn set_favorite(
         Ok(service) => service,
         Err(error) => {
             let _ = app.emit(
-                "set-favorite-failure",
+                "unfavorite-failure",
                 json!({ "status": "failed", "error": error.to_string() }),
             );
             return;
         }
     };
-    let favorite_data = favorite_service.favorite().await;
+    let result = favorite_service.unfavorite().await;
 
-    if !favorite_data.success {
-        let _ = app.emit("set-favorite-failure", json!({ "status": "failed" }));
+    info!("result {:?}", result);
+
+    if !result.success {
+        let _ = app.emit("unfavorite-failure", json!({ "status": "failed" }));
         return;
     }
 
-    let _ = app.emit("set-favorite-success", json!({ "status": "success" }));
+    let _ = app.emit("unfavorite-success", json!({ "status": "success" }));
 }
