+++ src-tauri/src/commands/assets.rs
@@ -1,4 +1,4 @@
-use crate::commands::auth_login::ensure_fresh_token;
+use crate::commands::auth::ensure_fresh_token;
 use crate::service::asset::{AssetQuery, AssetService};
 use log::{error, info};
 use serde::Deserialize;
