+++ src-tauri/src/commands/mod.rs
@@ -1,17 +1,17 @@
-pub mod auth_login;
-pub mod get_asset_detail;
-pub mod get_assets;
-pub mod get_config;
-pub mod get_connect_methods;
-pub mod get_setting;
-pub mod get_token;
-pub mod get_version_message;
-pub mod http_callback;
-pub mod list_system_fonts;
+pub mod asset_detail;
+pub mod asset_favorite;
+pub mod asset_rename;
+pub mod asset_unfavorite;
+pub mod assets;
+pub mod auth;
+pub mod client_launcher;
+pub mod config;
+pub mod config_update;
+pub mod connect_methods;
+pub mod connect_token;
+pub mod dev_http_server;
 pub mod logout;
-pub mod pull_up;
-pub mod rename_asset;
-pub mod set_favorite;
-pub mod unfavorite;
-pub mod update_config;
-pub mod window_controls;
+pub mod setting;
+pub mod system_fonts;
+pub mod version;
+pub mod window;
