+++ src-tauri/src/http/routes/mod.rs
@@ -1,2 +1,2 @@
-pub(crate) mod oauth_dev_callback;
+pub(crate) mod oauth_callback;
 pub(crate) mod recordings;
