+++ src-tauri/src/lib.rs
@@ -9,23 +9,23 @@ use crate::setup::apply_window_effects;
 use crate::setup::menu::{build_menu, handle_menu_event};
 use crate::setup::setup_tray;
 
-use crate::commands::auth_login::{auth_cancel, auth_login, handle_auth_callback, AuthFlowState};
-use crate::commands::get_asset_detail::get_asset_detail;
-use crate::commands::get_assets::get_assets;
-use crate::commands::get_config::get_config;
-use crate::commands::get_connect_methods::get_connect_methods;
-use crate::commands::get_setting::get_setting;
-use crate::commands::get_token::get_connect_token;
-use crate::commands::get_version_message::get_version_message;
-use crate::commands::http_callback::init_http_callback_server;
-use crate::commands::list_system_fonts::list_system_fonts;
+use crate::commands::asset_detail::get_asset_detail;
+use crate::commands::asset_favorite::set_favorite;
+use crate::commands::asset_rename::rename;
+use crate::commands::asset_unfavorite::unfavorite;
+use crate::commands::assets::get_assets;
+use crate::commands::auth::{auth_cancel, auth_login, handle_auth_callback, AuthFlowState};
+use crate::commands::client_launcher::pull_up;
+use crate::commands::config::get_config;
+use crate::commands::config_update::update_config_selection;
+use crate::commands::connect_methods::get_connect_methods;
+use crate::commands::connect_token::get_connect_token;
+use crate::commands::dev_http_server::init_http_callback_server;
 use crate::commands::logout::logout;
-use crate::commands::pull_up::pull_up;
-use crate::commands::rename_asset::rename;
-use crate::commands::set_favorite::set_favorite;
-use crate::commands::unfavorite::unfavorite;
-use crate::commands::update_config::update_config_selection;
-use crate::commands::window_controls::{close_window, minimize_window, toggle_maximize_window};
+use crate::commands::setting::get_setting;
+use crate::commands::system_fonts::list_system_fonts;
+use crate::commands::version::get_version_message;
+use crate::commands::window::{close_window, minimize_window, toggle_maximize_window};
 use crate::utils::is_auth_callback;
 
 use log::{error, info, warn};
