+++ src-tauri/src/service/asset_detail.rs
