+++ src-tauri/src/service/asset_favorite.rs
