+++ src-tauri/src/service/asset_rename.rs
